🔧(fret): fix critical type conversion errors and compilation issues

+++ src/fret/continuous.rs
@@ -42,7 +42,7 @@ impl ContinuousFretboard {
 
     /// Get the number of strings on this fretboard
     pub fn string_count(&self) -> usize {
-        self.config.string_count()
+        self.config.string_count() as usize
     }
 
     /// Get the tuning of a specific string
@@ -53,11 +53,11 @@ impl ContinuousFretboard {
     /// Calculate the tuning at a continuous position on a string
     /// Position 0.0 = open string, 1.0 = theoretical maximum position
     pub fn tuning_at_continuous_position(&self, position: &ContinuousPosition) -> Option<Tuning> {
-        if position.string >= self.string_count() {
+        if position.string >= self.string_count() as u32 {
             return None;
         }
 
-        let open_tuning = self.string_tuning(position.string)?;
+        let open_tuning = self.string_tuning(position.string as usize)?;
 
         // For continuous instruments, we use a logarithmic relationship
         // similar to fretted instruments but with continuous positions
@@ -114,7 +114,8 @@ impl ContinuousFretboard {
 
                     // Only include positions within practical playing range
                     if position_value <= 1.0 {
-                        positions.push(ContinuousPosition::new(string_index, position_value));
+                        positions
+                            .push(ContinuousPosition::new(string_index as u32, position_value));
                     }
                 }
             }
@@ -145,7 +146,7 @@ impl ContinuousFretboard {
 
     /// Validate that a position is within reasonable playing range
     pub fn is_position_playable(&self, position: &ContinuousPosition) -> bool {
-        position.string < self.string_count()
+        position.string < self.string_count() as u32
             && position.position >= 0.0
             && position.position <= 1.0
     }
@@ -198,7 +199,8 @@ impl Fretboard for ContinuousFretboard {
     fn get_range(&self) -> (Self::Position, Self::Position) {
         // For continuous instruments, the range is from position 0.0 to 1.0 on all strings
         let min_position = ContinuousPosition::new(0, 0.0);
-        let max_position = ContinuousPosition::new(self.string_count().saturating_sub(1), 1.0);
+        let max_position =
+            ContinuousPosition::new((self.string_count().saturating_sub(1)) as u32, 1.0);
         (min_position, max_position)
     }
 
@@ -214,7 +216,7 @@ impl Fretboard for ContinuousFretboard {
         for string_index in 0..self.string_count() {
             for i in 0..=100 {
                 let position_value = i as f32 / 100.0;
-                positions.push(ContinuousPosition::new(string_index, position_value));
+                positions.push(ContinuousPosition::new(string_index as u32, position_value));
             }
         }
 
@@ -508,8 +510,8 @@ mod tests {
 mod property_tests {
     use super::*;
     use crate::fret::presets::InstrumentPresets;
+    use crate::fret::types::{KeyboardPosition, StringedPosition};
     use crate::fret::{KeyboardFretboard, StringedFretboard};
-    use crate::fret::types::{StringedPosition, KeyboardPosition};
     use proptest::prelude::*;
 
     /// **Property 3: Position Type Support**
@@ -532,56 +534,56 @@ mod property_tests {
             // Test StringedFretboard with discrete positions
             let stringed_config = InstrumentPresets::guitar_standard();
             let stringed_fretboard = StringedFretboard::new(stringed_config).unwrap();
-            
+
             // Create a valid stringed position within the instrument's range
             let string_idx = string_index % stringed_fretboard.string_count();
             let fret_num = fret_number % stringed_fretboard.fret_count();
             let stringed_position = StringedPosition::new(string_idx, fret_num);
-            
+
             // Verify that StringedFretboard correctly handles StringedPosition
             let stringed_tuning = stringed_fretboard.tuning_at_position(&stringed_position);
-            prop_assert!(stringed_tuning.is_some(), 
+            prop_assert!(stringed_tuning.is_some(),
                         "StringedFretboard should handle StringedPosition correctly");
-            
+
             // Verify position validation works correctly
             prop_assert!(stringed_fretboard.is_position_valid(&stringed_position),
                         "Valid StringedPosition should be recognized as valid");
-            
+
             // Test ContinuousFretboard with continuous positions
             let continuous_config = create_violin_config();
             let continuous_fretboard = ContinuousFretboard::new(continuous_config).unwrap();
-            
+
             // Create a valid continuous position within the instrument's range
             let cont_string_idx = string_index % continuous_fretboard.string_count();
             let continuous_pos = ContinuousPosition::new(cont_string_idx, continuous_position);
-            
+
             // Verify that ContinuousFretboard correctly handles ContinuousPosition
             let continuous_tuning = continuous_fretboard.tuning_at_position(&continuous_pos);
             prop_assert!(continuous_tuning.is_some(),
                         "ContinuousFretboard should handle ContinuousPosition correctly");
-            
+
             // Verify position validation works correctly
             prop_assert!(continuous_fretboard.is_position_valid(&continuous_pos),
                         "Valid ContinuousPosition should be recognized as valid");
-            
+
             // Test KeyboardFretboard with keyboard positions
             let keyboard_config = InstrumentPresets::piano_88_key();
             let keyboard_fretboard = KeyboardFretboard::new(keyboard_config).unwrap();
-            
+
             // Create a valid keyboard position within the instrument's range
             let key_idx = key_index % keyboard_fretboard.key_count();
             let keyboard_position = KeyboardPosition::new(key_idx);
-            
+
             // Verify that KeyboardFretboard correctly handles KeyboardPosition
             let keyboard_tuning = keyboard_fretboard.tuning_at_position(&keyboard_position);
             prop_assert!(keyboard_tuning.is_some(),
                         "KeyboardFretboard should handle KeyboardPosition correctly");
-            
+
             // Verify position validation works correctly
             prop_assert!(keyboard_fretboard.is_position_valid(&keyboard_position),
                         "Valid KeyboardPosition should be recognized as valid");
         }
-        
+
         /// Test that each fretboard type correctly validates its position type
         #[test]
         fn prop_position_type_validation(
@@ -591,24 +593,24 @@ mod property_tests {
             // Test that continuous positions are properly bounded
             let continuous_config = create_violin_config();
             let continuous_fretboard = ContinuousFretboard::new(continuous_config).unwrap();
-            
+
             let string_idx = string_or_key_index % continuous_fretboard.string_count();
             let continuous_pos = ContinuousPosition::new(string_idx, position_value);
-            
+
             // All positions with valid string index and 0.0-1.0 range should be valid
             prop_assert!(continuous_fretboard.is_position_valid(&continuous_pos),
                         "ContinuousPosition with valid range should be valid");
-            
+
             // Test boundary conditions
             let boundary_pos_min = ContinuousPosition::new(string_idx, 0.0);
             let boundary_pos_max = ContinuousPosition::new(string_idx, 1.0);
-            
+
             prop_assert!(continuous_fretboard.is_position_valid(&boundary_pos_min),
                         "ContinuousPosition at 0.0 should be valid");
             prop_assert!(continuous_fretboard.is_position_valid(&boundary_pos_max),
                         "ContinuousPosition at 1.0 should be valid");
         }
-        
+
         /// Test that position-to-tuning calculations are consistent with position type
         #[test]
         fn prop_position_tuning_consistency(
@@ -617,26 +619,26 @@ mod property_tests {
         ) {
             let continuous_config = create_violin_config();
             let continuous_fretboard = ContinuousFretboard::new(continuous_config).unwrap();
-            
+
             let string_idx = string_index % continuous_fretboard.string_count();
-            
+
             // Test that higher positions on the same string produce higher pitches
             let lower_pos = ContinuousPosition::new(string_idx, position_value);
             let higher_pos = ContinuousPosition::new(string_idx, position_value + 0.1);
-            
+
             let lower_tuning = continuous_fretboard.tuning_at_position(&lower_pos);
             let higher_tuning = continuous_fretboard.tuning_at_position(&higher_pos);
-            
+
             if let (Some(lower), Some(higher)) = (lower_tuning, higher_tuning) {
                 prop_assert!(higher.number() >= lower.number(),
                            "Higher positions should produce higher or equal pitches");
             }
-            
+
             // Test that open string (position 0.0) matches the string tuning
             let open_pos = ContinuousPosition::new(string_idx, 0.0);
             let open_tuning = continuous_fretboard.tuning_at_position(&open_pos);
             let string_tuning = continuous_fretboard.string_tuning(string_idx);
-            
+
             if let (Some(open), Some(string)) = (open_tuning, string_tuning) {
                 prop_assert_eq!(open.class(), string.class(),
                               "Open position should match string tuning pitch class");
@@ -652,11 +654,11 @@ mod property_tests {
         StringedInstrumentConfig::new(
             vec![
                 Tuning::from_str("G3").unwrap(), // G string
-                Tuning::from_str("D4").unwrap(), // D string  
+                Tuning::from_str("D4").unwrap(), // D string
                 Tuning::from_str("A4").unwrap(), // A string
                 Tuning::from_str("E5").unwrap(), // E string
             ],
-            0, // No frets for continuous instruments
+            0,     // No frets for continuous instruments
             330.0, // Violin scale length in mm
             24.0,  // Nut width
             7.0,   // String spacing
